+++ jock/cli.py
@@ -1,28 +1,28 @@
 import click
 
-from jockcli.git import Git
+from jock.git import Git
 
 
 @click.group()
-def jock_cli():
+def main():
     pass
 
 
-@jock_cli.command()
+@main.command()
 @click.argument('repositories', type=str, nargs=-1)
 def clone(repositories):
     git = Git(repositories)
     git.clone()
 
 
-@jock_cli.command()
+@main.command()
 @click.argument('repositories', type=str, nargs=-1)
 def pull(repositories):
     git = Git(repositories)
     git.pull()
 
 
-@jock_cli.command()
+@main.command()
 @click.argument('repositories', type=str, nargs=-1)
 def fetch(repositories):
     git = Git(repositories)
@@ -30,4 +30,4 @@ def fetch(repositories):
 
 
 if __name__ == '__main__':
-    jock_cli()
+    main()
