--- jockcli/__main__.py
@@ -1,6 +1,0 @@
-import sys
-
-from jockcli.jockcli import jock_cli
-
-if __name__ == "__main__":
-    sys.exit(jock_cli())
