+++ setup.py
@@ -6,7 +6,7 @@ with open('README.md', 'r') as fh:
 
 setup(
     name='git-jock',
-    version='1.0.0b2',
+    version='1.0.0b3',
     author='Gavin Fenton',
     author_email='contact@gavinfenton.com',
     description='Git helper for multi-repository management',
